reafactor: redo pushController from component into method
